server/requirements.txt
Django>=5.0,<6.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
